Proposed redefinition of "Thing" vs. "Servient"
1. modified "links to domain-specific vocabulary" ("Thing Description"). It would be more accurate to say some of the metadata *is expressed* using domain-specific vocabularies.

2. proposed a new definition for "Thing" (main change). Details:

In the current version, the definitions for Thing and Servient are redundant: a Thing is "an entity that that natively implements WoT building blocks (e.g., Protocol Bindings or the Scripting API)." while a Servient is "an entity that implements WoT building blocks".

In the proposed definition, the servient is the software agent and the Thing is the physical entity (with possibly no computational power itself). The Internet of Things was first used to describe the phenomenon that non-computational objects get a digital representation via RFID tags. See Kevin Ashton's article about it: http://www.rfidjournal.com/articles/view?4986. From this perspective, a bottle, a can or any manufactured product can become a "Thing" as long as it has a digital tag (which is not even mandatory if we consider computer vision). To remain aligned with this idea, I think it is important that we define a WoT Thing in such an abstract way that it can be "any physical entity".