Add 6.5 Inter connection of servients
The building blocks is defined as common functions and structure derived from 3 types of the entities. The inter connection for basic configurations such as (1) application and device, (2) application, device and proxy is useful information for readers to understand the building blocks.  Therefore it should be described in the same section as "5. Building Blocks", and move the contents from the section "8. WoT Deployments".
On the other hands, "8. WoT Deployments" can focus to more actual conditions. The section "8.2. Device in a Local Network Controlled from a Cloud" and "8.3. Service-to-Service connections across multiple domains" are referred from the typical situation occurred in the commercial IoT systems.