New Figures
showing internal details and options for Thing implementation, with and without the scripting API.   Adds a new concept, the "Behavioral Agent" which has one of more execution contents inside it, said contexts OPTIONALLY supporting a WoT Runtime... or, alternatively, a native implementation.  The native implementation also shows the TD being provided externally by a Thing Directory (which will often, but not necessarily, be the case; certainly it will be the case when "retrofitting" an existing implementation you cannot modify).