Updates to overview images for press release
Add "Web of Things (WoT)" to legend for TD to allow image to stand alone if published out of context.
Note: SVG compressed with nano as requested.  However, font substitution still seems to take place when loaded into chrome.
PNG file seems fine (this is what is linked in the press release).