WIP: Explainer for Architecture 1.1
Initial placeholder for material to be refactored from TD 1.0 explainer.
See https://github.com/w3c/wot-thing-description/pull/1397
Will update this PR (e.g. with To Do list) after we discuss proposed refactoring plan.
Note that the name "Explainer11.md" is meant to be consistent with the use of versions elsewhere.
We may want to add a simple "Explainer.md" file as well here that just links to the one under the TD repo so people can find it.